requirements.txt
streamlit
opencv-python-headless
Pillow
numpy
pandas
ultralytics
requests
matplotlib        
scikit-learn 
opencv     